refactor: 게시글 관련 dto 디렉토리 개선. 불필요한 멤버 삭제
